Added PWM folder and Flight Data Folder,
Flight Data: First Flight data starts collecting data while rolling test flight on the runway proceeded. Real flight data (in air) does not start until somewhere between the first and seconds minute of data collection. 2nd flight starts as it should. Data collection starts when the aircraft is flat at the end of the runway. Aircraft starts takeoff soon after.

PWM folder: Pictures of PWM signals from the transmitter to the receiver are shown. This is the hardware from the "Storch." The inputs were from the aileron channel on the receiver.