Updated Code
Arduino Nano is now used, not the Uno.  Yellow light is commented out of the code since its use is transferred to the red LED. Button is also commented out of code as it is not needed in flight. Code will now start once it receives power. It will now first initialize everything and then wait a specific amount of time -- a 1 minute delay is used right now.